GOCBC-1072: Don't block in poller controller OnNewRouteConfig.
Motivation
----------
A scenario can occur where ForceHTTPPoller and OnNewRouteConfig
deadlock. This can happen when there are 2+ connections and one
successfully bootstraps whilst the other fails with bucket not
found. If cccp successfully gets a config at the same time as
the bucket not found connection returns up the stack to ForceHTTPPoller
then a deadlock can occur where ForceHTTPPoller is waiting for cccp
to complete but cccp is blocking by waiting for the controllerLock
lock in OnNewRouteConfig, which is already held by ForceHTTPPoller.

Changes
-------
OnNewRouteConfig already uses a goroutine when it starts the poller
controller again, so moved the go routine to emcompass the controller
lock. This way there is no blocking in that function.
Add a bucketConfigSeen atomic which is set in OnNewRouteConfig and
is checked at the start of ForceHTTPPoller and also after the
cccp done channel completes.

Change-Id: I0e522df20c722664a79e57e117b2935ace37794e